requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
gunicorn==20.1.0
python-dotenv==0.19.1
requests==2.26.0
numpy>=1.19.0
pandas==1.3.4
opencv-python>=4.5.0
ffmpeg-python==0.2.0
moviepy==1.0.3
openai==1.3.0
pydub==0.25.1
librosa==0.9.1
scikit-learn==1.0.1
Pillow>=8.0.0
pytest==6.2.5
pytest-cov==2.12.1
black==21.10b0
isort==5.9.3
flake8==4.0.1
mypy==0.910
SpeechRecognition==3.8.1
transformers==4.12.5
torch>=1.7.0
tensorflow==2.7.0
# Avatar Creation Module Dependencies
dlib>=19.21.0
face-alignment>=1.3.3
pytorch3d==0.7.2
kornia==0.6.9
stylegan3==0.0.1
mediapipe==0.10.0
open3d>=0.12.0
trimesh>=3.9.0
pyrender>=0.1.45
pyglet>=1.5.0
pywavefront>=1.3.0
# Enhanced Face Reconstruction Dependencies
scipy>=1.5.0
chumpy==0.70
tqdm>=4.50.0
einops==0.4.1
ninja==1.10.2
pyyaml==6.0
insightface==0.6.2
# Improved UV Mapping Dependencies
xatlas==0.0.7
# Enhanced StyleGAN3 Dependencies
matplotlib>=3.3.0
scikit-image>=0.17.0
lpips>=0.1.3
# Optional Dependencies
# Uncomment these if needed for specific features
# cupy>=8.0.0  # For GPU acceleration of certain operations
# tensorflow>=2.4.0  # For certain neural network models
# mxnet>=1.7.0  # For certain neural network models 